de_duplication/requirements.txt
requests
python-dotenv
aiokafka
redis
